Add ⌘-comma/⌃-comma shortcuts for opening the Settings dialog (#219)
Due to a limitation in Qt 6.6.0, the keyboard shortcut doesn't appear under the Mac's application menu. It's active and it works, but it's not shown in the menu as a valid keyboard shortcut. I'm putting this in the "better than nothing" category at this time. Ideally a future update to Qt will fix the problem without our having to change anything.