Quote current directory during update on Windows
Sometimes, when you install Windows 11 and use a Microsoft account it decides that your username and therefore your home directory in C:\Users\ should be your full name with a space in it. The update script quotes the source directory path but not the destination, so the robocopy command fails if you have a space anywhere in the latter path.