Default Text Editor // Open Mod Folder in Text Editor Action (#1269)
- add a setting for a default text editor 
- This is on its own tab for now, "Other External Tools". Very willing
to move it if someone thinks it fits in better with an existing tab.
- add a right click action to open a mod folder with the default text
editor, if one is set
  - Defaults to empty, so the context action doesn't appear by default
- add a right click action to open a file path in the default text
editor if the selected line in the Player Log Viewer has one.
- Not yet integrated with the file search tab
- file search: Already has existing editor tools. I see two paths here -
we can remove the existing shortcuts, or I can just use the text editor
setting to drive the "default" case and leave the fixed options of
Notepad++ and code in place.

---------

Signed-off-by: Dave Lee <dave@gray101.com>