Fix import collection not importing all mods (#1476)
Problem: Import collection feature wasn't importing all mods. This was
because some records in steam workshop database are missing package id
field.

I changed it so that the tool tries to get package id in three ways

* with steam workshop database
* from internal metadata (picking packageId from workshop folder)
* in case the mod in the workshop folder is missing publishedfileid, it
will assume that the publishedfileid is the mod's folder name

If package id still can't be found, it tries to access the mod's steam
workshop page and see if the mod is available.

It will now only include actual mods (not collections for instance),
i.e. only items with "filetype"=0